Split focused test support helpers
Move polling utilities and PDF text extraction out of the broad core test-support module into purpose-named support modules with direct imports.

Remove unused maintenance feature timer and fetch helper leftovers so the shared UI support file stops carrying dead setup paths.
