+++ apps/server/tests/adapters/pdf/test_report_pdf_content_rendering.py
@@ -7,7 +7,7 @@ from io import BytesIO
 
 from _paths import SERVER_ROOT
 from pypdf import PdfReader
-from test_support.core import extract_pdf_text
+from test_support.pdf import extract_pdf_text
 
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
 from vibesensor.shared.boundaries.reporting.document import (
