+++ apps/server/tests/adapters/pdf/test_report_pdf_diagram_visual_audit.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from test_support.core import extract_pdf_text
+from test_support.pdf import extract_pdf_text
 
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
 from vibesensor.domain import LocationIntensitySummary
