+++ apps/server/tests/adapters/pdf/test_report_pdf_usefulness_layout.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from io import BytesIO
 
 from pypdf import PdfReader
-from test_support.core import extract_pdf_text
+from test_support.pdf import extract_pdf_text
 
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
 from vibesensor.shared.boundaries.reporting.document import (
