+++ apps/server/tests/use_cases/run/test_metrics_logger_lifecycle.py
@@ -6,8 +6,8 @@ import logging
 from pathlib import Path
 
 import pytest
-from test_support.core import wait_until
 from test_support.persisted_analysis import make_persisted_analysis
+from test_support.polling import wait_until
 
 from vibesensor.adapters.persistence.history_db import create_history_persistence_adapters
